prototype/training/requirements.txt
transformers>=4.36.0
torch>=2.0.0
datasets>=2.16.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
accelerate>=0.25.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
